+++ R/createFromLm.R
@@ -35,9 +35,16 @@ createFromLm <- function(lmObj, outcome_name = NULL) {
         stop("An intercept term should be included in the model.")
     }
 
-    # Calculate Xty and XtWX
+    # Calculate Xty and XtX
     Xty <- t(X) %*% y
     XtX <- t(X) %*% X
+
+    # Calculate the sample size
+    n <- nrow(X)
+    # Calculate the mean of the outcome variable
+    y_mean <- mean(y, na.rm = TRUE)
+    # Calculate the sum of squared responses
+    yty <- as.numeric(crossprod(y))
 
     # Get the outcome name (if not provided)
     if (is.null(outcome_name)) {
@@ -48,5 +55,10 @@ createFromLm <- function(lmObj, outcome_name = NULL) {
     colnames(Xty) <- outcome_name
 
     # Create ftmlm object
-    ftmlm(XtX = XtX, Xty = Xty)
+    ftmlm(XtX = XtX,
+          Xty = Xty,
+          s = 0,
+          n = n,
+          yty = yty,
+          y_mean = y_mean)
 }